fixed logic

+++ scripts/IKIWIKI/40-remove_index_html.sh
@@ -1,4 +1,6 @@
 #! /bin/sh
 
 # gets in the way of wordpress's index.php
-$ROOTCMD rm /var/www/index.htm
+if $ROOTCMD -e test /var/www/index.htm; then
+    $ROOTCMD rm /var/www/index.htm
+fi
